requirements.txt
channels==3.0.4
channels-redis==3.3.0
Django==3.2.6
django-allauth==0.45.0
django-apscheduler==0.6.0
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-constance==2.8.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-phonenumber-field==5.2.0
django-picklefield==3.0.1
django-redis==5.0.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-timezone-field==4.2.1
vonage==2.5.5
celery==5.1.2
requests==2.25.1

APScheduler==3.7.0
asgiref~=3.4.1
atomicwrites==1.4.0
attrs~=20.3.0
certifi~=2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml~=0.6.0
flake8
factory_boy
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
oauthlib==3.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest~=6.2.2
pytest-django==4.4.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis
requests==2.25.1
requests-oauthlib==1.3.0
six~=1.15.0
sorl-thumbnail==12.7.0
Sphinx
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3~=1.26.4
psycopg2
djangorestframework==3.12.4